stop watch
The code implements a digital clock on an AVR microcontroller with a seven segment display. Timer1 is configured in CTC mode to generate interrupts at 1 second interval using prescaling.

Time values are stored in individual variables and incremented in the timer interrupt. Interrupts are also used to pause, resume and reset the clock on external button presses.

The main() function initializes ports, enables interrupts and runs in a loop. Time values are multiplexed on the display by activating one segment pin at a time along with relevant port values.

Functions are used to configure timers, interrupts and drive the display in a modular way. Register macros and bit level operations are used efficiently.